# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "{{.Description}}",
        "title": "{{.Title}}",
        "contact": {},
        "license": {},
        "version": "{{.Version}}"
    },
    "host": "{{.Host}}",
    "basePath": "{{.BasePath}}",
    "paths": {
...
        "/v2/transactions": {
            "post": {
                "description": "发起批量交易",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v2-general"
                ],
                "summary": "发起批量交易",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/bean.SignedBatchTx"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/bean.PublicResp"
                        }
                    }
                }
            }
        },
...
        "/v3/accounts/{address}/transactions": {
            "get": {
                "description": "根据账户地址查询交易记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v3-query"
                ],
                "summary": "根据账户地址查询交易记录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "账户地址",
                        "name": "address",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "开始时间戳",
                        "name": "begin",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "结束时间戳",
                        "name": "end",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "游标",
                        "name": "cursor",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "限制",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序(ASC/DESC)",
                        "name": "order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/database.V3Transaction"
                            }
                        }
                    }
                }
            }
        },
...
        "/v3/ext/price/{symbol}": {
            "get": {
                "description": "查询币种价格：币种-USDT交易对",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v3-extension"
                ],
                "summary": "查询币种价格",
                "parameters": [
                    {
                        "type": "string",
                        "description": "币种",
                        "name": "symbol",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/bean.Ticker"
                        }
                    }
                }
            }
        },
...
        "/v3/ledgers/{height}/transactions": {
            "get": {
                "description": "根据区块高度查询交易记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v3-query"
                ],
                "summary": "根据区块高度查询交易记录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "区块高度",
                        "name": "height",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "开始时间戳",
                        "name": "begin",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "结束时间戳",
                        "name": "end",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "游标",
                        "name": "cursor",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "限制",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序(ASC/DESC)",
                        "name": "order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/database.V3Transaction"
                            }
                        }
                    }
                }
            }
        },
...
        "/v3/transactions": {
            "get": {
                "description": "根据所有交易记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v3-query"
                ],
                "summary": "根据所有交易记录",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "开始时间戳",
                        "name": "begin",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "结束时间戳",
                        "name": "end",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "游标",
                        "name": "cursor",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "限制",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序(ASC/DESC)",
                        "name": "order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/database.V3Transaction"
                            }
                        }
                    }
                }
            }
        },
        "/v3/transactions/{txhash}": {
            "get": {
                "description": "根据txhash查询交易记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v3-query"
                ],
                "summary": "根据txhash查询交易记录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "交易hash",
                        "name": "txhash",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/database.V3Transaction"
                            }
                        }
                    }
                }
            }
        },
...
        "/v3/voters/{address}": {
            "get": {
                "description": "查询投票者信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "v3-dpos"
                ],
                "summary": "查询投票者信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "地址",
                        "name": "address",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/types.BeanVoter"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "bean.Operation": {
            "type": "object",
            "properties": {
                "to": {
                    "description": "交易接受方地址，可以是普通用户地址、合约地址、节点账户地址",
                    "type": "string"
                },
                "value": {
                    "description": "交易金额",
                    "type": "string"
                }
            }
        },
...
        "bean.PublicResp": {
            "type": "object",
            "properties": {
                "isSuccess": {
                    "description": "是否成功",
                    "type": "boolean"
                },
                "message": {
                    "description": "消息提醒",
                    "type": "string"
                },
                "result": {
                    "description": "数据对象",
                    "type": "object"
                }
            }
        },
...
        "bean.SignedBatchTx": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "description": "时间戳，可选字段，秒/毫秒均可",
                    "type": "integer"
                },
                "gasLimit": {
                    "description": "gas限额",
                    "type": "integer"
                },
                "gasPrice": {
                    "description": "gas价格，至少为1",
                    "type": "string"
                },
                "memo": {
                    "description": "备注，必须\u003c256字节",
                    "type": "string"
                },
                "mode": {
                    "description": "模式:0-default/commit 1-async 2-sync",
                    "type": "integer"
                },
                "nonce": {
                    "description": "用户nonce，每次交易前从链上获取，每次交易nonce+1",
                    "type": "integer"
                },
                "operations": {
                    "description": "交易列表，数量不可大于10000笔",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/bean.Operation"
                    }
                },
                "sender": {
                    "description": "交易发起者公钥",
                    "type": "string"
                },
                "signature": {
                    "description": "交易签名的hex字符串",
                    "type": "string"
                }
            }
        },
...
        "bean.Ticker": {
            "type": "object",
            "properties": {
                "amount": {
                    "description": "交易金额",
                    "type": "string"
                },
                "buy": {
                    "description": "买入价",
                    "type": "number"
                },
                "high": {
                    "description": "最高价",
                    "type": "string"
                },
                "last": {
                    "description": "最新成交价",
                    "type": "number"
                },
                "low": {
                    "description": "最低价",
                    "type": "string"
                },
                "rose": {
                    "description": "振幅",
                    "type": "string"
                },
                "sell": {
                    "description": "卖出价",
                    "type": "number"
                },
                "time": {
                    "description": "时间戳 ms",
                    "type": "integer"
                },
                "vol": {
                    "description": "成交量",
                    "type": "string"
                }
            }
        },
...
        "database.V3Transaction": {
            "type": "object",
            "properties": {
                "codei": {
                    "description": "失败代码",
                    "type": "integer"
                },
                "codes": {
                    "description": "失败原因",
                    "type": "string"
                },
                "createdAt": {
                    "description": "区块时间",
                    "type": "string"
                },
                "events": {
                    "description": "交易事件",
                    "type": "string"
                },
                "gasLimit": {
                    "description": "gas限额",
                    "type": "integer"
                },
                "gasPrice": {
                    "description": "gas价格",
                    "type": "string"
                },
                "gasUsed": {
                    "description": "gas使用量",
                    "type": "integer"
                },
                "hash": {
                    "description": "交易hash",
                    "type": "string"
                },
                "height": {
                    "description": "区块高度",
                    "type": "integer"
                },
                "id": {
                    "description": "数据库自增id",
                    "type": "integer"
                },
                "memo": {
                    "description": "备注",
                    "type": "string"
                },
                "nonce": {
                    "description": "交易发起者nonce",
                    "type": "integer"
                },
                "payload": {
                    "description": "负载",
                    "type": "string"
                },
                "receiver": {
                    "description": "交易接受者地址",
                    "type": "string"
                },
                "sender": {
                    "description": "交易发起者地址",
                    "type": "string"
                },
                "typei": {
                    "description": "交易类型",
                    "type": "integer"
                },
                "types": {
                    "description": "交易类型",
                    "type": "string"
                },
                "value": {
                    "description": "交易金额",
                    "type": "string"
                }
            }
        },
...
        "types.BeanVoter": {
            "type": "object",
            "properties": {
                "address": {
                    "description": "选民地址",
                    "type": "string"
                },
                "balance": {
                    "description": "累积收益余额",
                    "type": "string"
                },
                "burned": {
                    "description": "已燃烧",
                    "type": "string"
                },
                "mortgaged": {
                    "description": "已抵押",
                    "type": "string"
                },
                "redeem": {
                    "description": "赎回标志",
                    "type": "boolean"
                },
                "toMortgage": {
                    "description": "待抵押",
                    "type": "string"
                },
                "validator": {
                    "description": "节点地址",
                    "type": "string"
                }
            }
        },
...
    }
}